+++ ferrodec-decimal64/src/lib.rs
@@ -28,7 +28,13 @@ mod classify_types;
 mod cmp;
 mod convert;
 mod decimal;
+#[cfg(feature = "num-traits")]
+mod num_traits_impls;
 mod ops;
+#[cfg(feature = "ops")]
+mod ops_traits;
+#[cfg(feature = "serde")]
+mod serde_impls;
 mod status;
 #[cfg(kani)]
 mod verify;
@@ -37,4 +43,6 @@ pub use classify_types::IeeeClass;
 #[cfg(feature = "fmt")]
 pub use convert::{Engineering, ParseDecimalError};
 pub use decimal::{Decimal64, Decimal64BuildError};
+#[cfg(feature = "serde")]
+pub use serde_impls::serde_bid;
 pub use status::{RoundingMode, Status};